feat:提交返点比率模板相关接口

+++ src/main/java/com/dx/facade/member/param/UpdateRebateRateTemplateMappingParamDTO.java
@@ -38,7 +38,7 @@ public class UpdateRebateRateTemplateMappingParamDTO implements Serializable {
     private Long userId;
     
     @ApiModelProperty("用户类型 0：会员  1：代理")
-    private Long userType;
+    private Integer userType;
     
     @ApiModelProperty("总代Id")
     private Long topProxyId;
